Rename "jinja" package to "file_generators"
