+++ scargo/file_generators/cpp_gen.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Any, Dict
 
 from scargo.config import Config
-from scargo.jinja.base_gen import create_file_from_template
+from scargo.file_generators.base_gen import create_file_from_template
 
 
 class _CppTemplateGen:
