+++ scargo/file_generators/templates/.gitlab-ci.yml.j2
