+++ scargo/file_generators/templates/CMakeLists.txt.j2
