+++ scargo/file_generators/templates/cpp/cmake-src-esp32.j2
