+++ scargo/file_generators/templates/cpp/cmake-src-stm32.j2
