+++ scargo/file_generators/templates/cpp/cmake-src-x86.j2
