+++ scargo/file_generators/templates/cpp/lib.cpp.j2
