+++ scargo/file_generators/templates/cpp/lib.h.j2
