+++ scargo/file_generators/templates/cpp/main.cpp.j2
