+++ scargo/file_generators/templates/docker/Dockerfile-stm32.j2
