+++ scargo/file_generators/templates/docker/Dockerfile-x86.j2
