+++ scargo/file_generators/templates/docker/Dockerfile.j2
