+++ scargo/file_generators/templates/docker/devcontainer.json.j2
