+++ scargo/file_generators/templates/docker/docker-compose.yaml.j2
