+++ scargo/file_generators/templates/docker/stm32.cfg.j2
