+++ scargo/file_generators/templates/esp32.cmake.j2
