+++ scargo/file_generators/templates/mock/.clang-format
