+++ scargo/file_generators/templates/mock/CMakeLists.txt
