+++ scargo/file_generators/templates/mock/class_interface.h.j2
