+++ scargo/file_generators/templates/mock/class_mock.cpp.j2
