+++ scargo/file_generators/templates/mock/class_mock.h.j2
