+++ scargo/file_generators/templates/project.cmake.j2
