+++ scargo/file_generators/templates/scargo.toml.j2
