+++ scargo/file_generators/templates/stm32.cmake.j2
