+++ scargo/file_generators/templates/tests/CMakeLists-mocks.txt.j2
