+++ scargo/file_generators/templates/tests/static_mock/CMakeLists.txt
