+++ scargo/file_generators/templates/tests/static_mock/static_mock.h
