+++ scargo/file_generators/templates/x86.cmake.j2
